requirements.txt
opencv-python-headless
langchain==0.1.6
langchain-openai==0.0.6
langchain-core==0.1.22
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit~=1.31.1
faiss-cpu==1.7.4
altair<5
openai~=1.12.0
tiktoken
bs4
pexels-api-py
python-pptx
opencv-python~=4.9.0.80

requests~=2.31.0
pytube~=15.0.0